+++ scripts/z_nomenclature.py
@@ -210,5 +210,7 @@ def kew_query(occs, working_directory, verbose=True, debugging=False):
 # tmp1 = kew_query(debug_occs, 'Users/serafin/Sync/1_Annonaceae/share_DB_WIP/', verbose = True)
 # print('TMP!:', tmp1)
 
+# df=pd.read_csv('/Users/serafin/Sync/1_Annonaceae/G_Am_GLOBAL_Distr/4_post-processing/9_tmp_data/20231018_tmp-2.csv', sep = ';')
+# df_checked = kew_query(df, 'Users/serafin/Sync/1_Annonaceae/G_AM_GLOBAL_Distr/4_post-processing/', verbose = True)
 
-#
+# #